tests: fix deleted files unit test
